requirements.txt
pandas
pyodbc
mysql-connector-python
azure-storage-blob
sqlalchemy
chardet 
